add wrapper creator for basis functions to make them take the full model parameters
